requirements.txt
fastapi
uvicorn
pdfplumber
python-multipart
pytesseract
Pillow